Refactor webview
