+++ app/src/lib/editor/engine/overlay/index.ts
@@ -1,3 +1,4 @@
+import { getRelativeOffset } from '@/lib/utils';
 import { ClickRect, EditRect, HoverRect, ParentRect } from './rect';
 import { querySelectorCommand } from '/common/helpers';
 
@@ -38,18 +39,6 @@ export class OverlayManager {
         this.clear = this.clear.bind(this);
     };
 
-    // Helper function to calculate the relative offset to a common ancestor
-    getRelativeOffset(element: HTMLElement, ancestor: HTMLElement) {
-        let top = 0,
-            left = 0;
-        while (element && element !== ancestor) {
-            top += element.offsetTop || 0;
-            left += element.offsetLeft || 0;
-            element = element.offsetParent as HTMLElement;
-        }
-        return { top, left };
-    }
-
     getBoundingRect(selector: string, sourceWebview: Electron.WebviewTag) {
         return sourceWebview.executeJavaScript(
             `${querySelectorCommand(selector)}.getBoundingClientRect().toJSON()`,
@@ -66,9 +55,9 @@ export class OverlayManager {
 
     adaptRectFromSourceElement(rect: DOMRect, sourceWebview: Electron.WebviewTag) {
         const commonAncestor = this.overlayContainer?.parentElement as HTMLElement;
-        const sourceOffset = this.getRelativeOffset(sourceWebview, commonAncestor);
+        const sourceOffset = getRelativeOffset(sourceWebview, commonAncestor);
         const overlayOffset = this.overlayContainer
-            ? this.getRelativeOffset(this.overlayContainer, commonAncestor)
+            ? getRelativeOffset(this.overlayContainer, commonAncestor)
             : { top: 0, left: 0 };
         const adjustedRect = {
             ...rect,
