+++ app/src/lib/utils.ts
@@ -4,3 +4,15 @@ import { twMerge } from 'tailwind-merge';
 export function cn(...inputs: ClassValue[]) {
     return twMerge(clsx(inputs));
 }
+
+// Helper function to calculate the relative offset to a common ancestor
+export function getRelativeOffset(element: HTMLElement, ancestor: HTMLElement) {
+    let top = 0,
+        left = 0;
+    while (element && element !== ancestor) {
+        top += element.offsetTop || 0;
+        left += element.offsetLeft || 0;
+        element = element.offsetParent as HTMLElement;
+    }
+    return { top, left };
+}
