+++ app/src/routes/project/WebviewArea/index.tsx
@@ -2,8 +2,8 @@ import { WebviewMessageBridge } from '@/lib/editor/messageBridge';
 import { WebviewMetadata } from '@/lib/models';
 import { nanoid } from 'nanoid';
 import { useEditorEngine } from '..';
+import Frame from './Frame';
 import Overlay from './Overlay';
-import Webview from './Webview';
 
 function WebviewArea() {
     const editorEngine = useEditorEngine();
@@ -15,11 +15,12 @@ function WebviewArea() {
             src: ' http://localhost:3000/',
         },
     ];
+
     return (
         <Overlay>
             <div className="grid grid-flow-col gap-72">
                 {webviews.map((metadata, index) => (
-                    <Webview key={index} metadata={metadata} messageBridge={messageBridge} />
+                    <Frame key={index} metadata={metadata} messageBridge={messageBridge} />
                 ))}
             </div>
         </Overlay>
